code of Stats component splitted in many components
